+++ example/samples/urls.py
@@ -5,6 +5,7 @@ from django.utils.translation import ugettext_lazy as _
 from django.views.generic import TemplateView
 from django.shortcuts import redirect
 from . import views
+from backend.models import Artist
 
 app_name = 'samples'
 
@@ -29,6 +30,18 @@ urlpatterns = [
     path('form-submission/',
         TemplateView.as_view(template_name="samples/form_submission.html"),
         name="form-submission"),
+    path('edit-a-django-model-basic/',
+        TemplateView.as_view(
+            template_name="samples/edit_a_django_model_basic.html",
+            get_context_data=lambda: {'artists': Artist.objects.all()},
+        ),
+        name="edit-a-django-model-basic"),
+    path('edit-a-django-model-revised/',
+        TemplateView.as_view(
+            template_name="samples/edit_a_django_model_revised.html",
+            get_context_data=lambda: {'artists': Artist.objects.all()},
+        ),
+        name="edit-a-django-model-revised"),
     path('form-advanced/',
         TemplateView.as_view(template_name="samples/form_advanced.html"),
         name="form-advanced"),
@@ -38,6 +51,13 @@ urlpatterns = [
     path('advanced-form-validation', views.advanced_form_validation, name="advanced-form-validation"),
     path('form-validation-with-feedback', views.form_validation_with_feedback, name="form-validation-with-feedback"),
     path('simple-form-validation-with-addon', views.simple_form_validation_with_addon, name="simple-form-validation-with-addon"),
+
+    path('artist/add-basic/', views.add_artist, name="artist-add-basic"),
+    path('artist/<uuid:pk>/change-basic/', views.update_artist, name="artist-change-basic"),
+
+    path('artist/add/', views.edit_artist, name="artist-add"),
+    path('artist/<uuid:pk>/change/', views.edit_artist, name="artist-change"),
+    path('artist/<uuid:pk>/delete/', views.delete_artist, name="artist-delete"),
 
     path('new-track', views.new_track, name="new-track"),
     path('new-track-ex', views.new_track_ex, name="new-track-ex"),